+++ src/platform_impl/android/mod.rs
@@ -862,6 +862,10 @@ impl Window {
     pub fn theme(&self) -> Option<Theme> {
         None
     }
+
+    pub fn title(&self) -> String {
+        String::new()
+    }
 }
 
 #[derive(Default, Clone, Debug)]