+++ src/platform_impl/linux/x11/window.rs
@@ -1561,4 +1561,9 @@ impl UnownedWindow {
     pub fn theme(&self) -> Option<Theme> {
         None
     }
+
+    #[inline]
+    pub fn title(&self) -> String {
+        String::new()
+    }
 }