+++ src/platform_impl/web/window.rs
@@ -387,6 +387,11 @@ impl Window {
     pub fn theme(&self) -> Option<Theme> {
         None
     }
+
+    #[inline]
+    pub fn title(&self) -> String {
+        String::new()
+    }
 }
 
 impl Drop for Window {